services/task_service/pyproject.toml
[project]
name = "task-service"
version = "0.1.0"
description = "Task Manager API"
authors = [
    {name = "danya",email = "reactive722@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pymongo (>=4.11.2,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "redis",
    "requests"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
